+++ scripts/install.sh
@@ -212,11 +212,11 @@ mkdir --parent "$WORKSPACE/test/output/html"
 
 PATH_DIR_PUBLIC=$(getvar PATH_DIR_PUBLIC)
 
-rm "$PATH_DIR_PUBLIC/$(basename "$DIR_PUBLIC_OUTPUT")"
-ln -s "$(getvar DIR_PUBLIC_OUTPUT)" "$(getvar PATH_DIR_PUBLIC)/."
+rm "$PATH_DIR_PUBLIC/$(basename "$PATH_OUTPUT_LOCAL")"
+ln -s "$(getvar PATH_OUTPUT_LOCAL)" "$(getvar PATH_DIR_PUBLIC)/."
 
-rm "$PATH_DIR_PUBLIC/$(basename "$DIR_HTML_INPUT")"
-ln -s "$(getvar DIR_HTML_INPUT)" "$(getvar PATH_DIR_PUBLIC)/."
+rm "$PATH_DIR_PUBLIC/$(basename "$PATH_INPUT_LOCAL")"
+ln -s "$(getvar PATH_INPUT_LOCAL)" "$(getvar PATH_DIR_PUBLIC)/."
 
 # Use netlify redirect instead
 #   Add index.html